backend/requirements.txt
fastapi
sqlalchemy
mysql-connector-python
pydantic
uvicorn
passlib[bcrypt]==1.7.4
python-jose
python-dotenv
bcrypt==4.0.1
